♻️ rename files and improve invalid data handling
